First version of sender part of POV.
Reads frames from UART0 at 2mbps. Syncs on delay in data stream of >0.2s.

Transmits all frames read over nRF24L01+ to the main POV controller.

(One problem with this is that the built-in serial<->uart on the Tiva
Launchpad does not seem to be able to run at 2mbps, I get only a bit
more than 600mbps. So frame rate will be much too low. Hopefully this
is enough to at least get something running. Then possible solutions
could be to read from an SD card instead, or implement a custon USB
protocol that allows for sufficient frame speed.)