Change pov_sender to receive data directly from USB.
This way no separate usb-to-serial converter is needed, and we also
avoid the bottlenecks that some of those have, preventing full 25 FPS
operation.

There is still a bug in that if the internal buffer fills up, we can
stall USB transfer and fail to re-start it again. With an 8kB buffer
and framerate_control.pl to limit the transfer to the nominam 25FPS,
this does not seem to actually happen though it would be nice to get
it fixed anyway.