First working flash over wireless.
This version successfully can flash a test app on the target over the wireless.
(This part is the sender side, working as a USB-connected programmer on the PC).

There is still some glitches when connecting to the bootloader, if the
bootloader is not already running when wireless-flash.pl starts. This
needs to be fixed later. For now, to flash, reset the target (so it
enters the bootloader) and start the programmer while the bootloader
is still listening for packets.