Integrate brushless motor-controller code in pov_sender.
The basic code from blmotor_control is now included. It runs
completely in timer PWM interrupts, at highest priority to avoid any
mechanical glitches.

The code was optimised a bit. The motor-controller part now spends
around 50% of available CPU time (most of this is the computation of 3
sinf() calls every 1/50000 seconds).

Makefile adjusted to support hardware floating point; for this, the
StellarisWare code was recompiled with -ffloat-abi=hard (fixing the
FPU define in `makedefs').

Linker script was adjusted to avoid a strange error about overlapping
section .ARM.exidx.

So now should be able to use the same board for driving the motor and
for wireless flashing / communicating with the LED-torus.